requirements.txt
asgiref==3.7.2
Django==5.0.3
gitdb==4.0.11
GitPython==3.1.42
smmap==5.0.1
sqlparse==0.4.4
typing_extensions==4.10.0
